Work Awards activity moved to the more general place
