+++ app/src/main/kotlin/ru/fantlab/android/ui/modules/awards/item/ItemAwardsMvp.kt
@@ -1,10 +1,10 @@
-package ru.fantlab.android.ui.modules.work.awards
+package ru.fantlab.android.ui.modules.awards.item
 
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import ru.fantlab.android.data.dao.model.Awards
 import ru.fantlab.android.ui.base.mvp.BaseMvp
 
-interface WorkAwardsMvp {
+interface ItemAwardsMvp {
 
 	interface View : BaseMvp.View,
 			SwipeRefreshLayout.OnRefreshListener,
